% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAOU.R
\name{GetAOU}
\alias{GetAOU}
\title{GetAOU}
\usage{
GetAOU(alpha = NULL, common = NULL)
}
\arguments{
\item{alpha}{Data frame containing the (buffered) count data for the focal species}

\item{common}{Integer vector containing the bioclim variables of interest}
}
\value{
The four digit numeric code for the species of interest
}
\description{
Get numeric AOU species code using common name or four letter alpha code
}
