% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBioVars.R
\name{GetBioVars}
\alias{GetBioVars}
\title{GetBioVars}
\usage{
GetBioVars(alpha, index = c(1, 2, 8, 12, 18), path, ind_name = c("tmp",
  "dtr", "Twet", "Prec", "Pwarm"))
}
\arguments{
\item{alpha}{alpha code for species of interest}

\item{index}{Integer vector containing the bioclim variables of interest}

\item{ind_name}{Character vector containing abbreviated name of the bioclim variables of interest}
}
\value{
A .csv files containing the following fields:

routeID The unique 8 digit route ID for each route

Latitude The latitude for the route

Longitude The longitude for the route

Ind_Year The value for biovar[Ind] in year Year
}
\description{
Extract scaled annual bioclim variable values for each BBS route
}
