% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIndices.R
\name{GetIndices}
\alias{GetIndices}
\title{GetIndices}
\usage{
GetIndices(spp = NULL, alpha = NULL, path)
}
\arguments{
\item{alpha}{Vector of alpha codes for species of interest}
}
\value{
Data frame containing annual estimates of the following indices:

avg.psi = proportion of area occupied (i.e., range size)

s.lat = southern range limit

s.core = southern core range limit

n.lat = northern range limit

n.core = northern core range limit

w.lat = western range limit

w.core = western core range limit

e.lat = eastern range limit

e.core = eastern core range limit

avg.lat = occupancy-weighted mean breeding latitude

avg.lon = occupancy-weighted mean breeding longitude
}
\description{
Estimate indices of range dynamics from annual occupancy estimates
}
