% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInits.R
\name{GetInits}
\alias{GetInits}
\title{GetInits}
\usage{
GetInits(alpha = NULL, path)
}
\arguments{
\item{alpha}{four letter alpha code for species of interest; if NULL, runs Presence model for all species in parallel}
}
\description{
Estimate model coefficients in Presence; save output for mu/se values & inits
}
