% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOcc.R
\name{GetOccProb}
\alias{GetOccProb}
\title{GetOccProb}
\usage{
GetOccProb(alpha = NULL, path)
}
\arguments{
\item{alpha}{Four letter alpha code for species of interest}

\item{spp}{Vector containing alpha codes for all species of interest}
}
\description{
Predict annual probability of occupancy using parameter estimates from top model
}
