% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc.R
\name{obs_h}
\alias{obs_h}
\title{obs_h}
\usage{
obs_h(h, year, sim = FALSE)
}
\arguments{
\item{h}{detection history matrix}

\item{year}{Year for with observed counts needed}
}
\value{
List containing observed counts for each detection history & index of NA counts
}
\description{
Count observed number of routes with each detection history
}
