% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OccSummary.R
\name{OccSummary}
\alias{OccSummary}
\title{OccSummary}
\usage{
OccSummary(spp = NULL, alpha = NULL, path)
}
\arguments{
\item{spp}{Vector containing alpha codes for all species of interest}

\item{alpha}{Four letter alpha code for species of interest}
}
\value{
Data frame containing annual estimates of the following indices:
}
\description{
Estimate mean & CI psi for all raster cells
}
