% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc.R
\name{ppc}
\alias{ppc}
\title{ppc}
\usage{
ppc(spp = NULL, alpha = NULL, path)
}
\arguments{
\item{alpha}{alpha code for species of interest}
}
\value{
Bayesian p-value
}
\description{
Perform posterior predictive check on fitted model
}
