% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMod.R
\name{RunMod}
\alias{RunMod}
\title{RunMod}
\usage{
RunMod(spp = NULL, alpha = NULL, path, nI = 50000, nB = 10000,
  nC = 2, nT = 50, cores = 50, Parallel = FALSE)
}
\arguments{
\item{alpha}{alpha code for species of interest; if NULL, runs model in parallel for all species}
}
\description{
Run model
}
