% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pao.R
\name{write_pao}
\alias{write_pao}
\title{write_pao}
\usage{
write_pao(alpha)
}
\arguments{
\item{alpha}{Four letter alpha code for the species of interest}
}
\value{
A .pao file containing the detection histories, covariates, and summary information to input into Presence
}
\description{
Write occupancy and climate data for program Presence
}
